Split board souce file
