+++ src/AiBoard.cpp
@@ -1,15 +1,4 @@
-#ifndef AI_BOARD_H_
-#define AI_BOARD_H_
-
-#include "BoardBase.hpp"
-
-
-class AiBoard : public BoardBase {
-public:
-  AiBoard(const BoardBase& src);
-  const AiBoard& operator=(const AiBoard& src) = delete;
-  double get_status_score() const;
-};
+#include "../include/AiBoard.hpp"
 
 AiBoard::AiBoard(const BoardBase& src)
   : BoardBase(src)
@@ -56,5 +45,3 @@ double AiBoard::get_status_score() const
   }
   return status_score;
 }
-
-#endif // AI_BOARD_H_
