+++ src/GameBoard.cpp
@@ -1,22 +1,4 @@
-#ifndef GAME_BOARD_H_
-#define GAME_BOARD_H_
-
-#include <stdexcept>
-
-#include "BoardBase.hpp"
-
-class GameBoard : public BoardBase {
-public:
-  GameBoard(const BoardBase::Position size = BoardBase::Position{8, 8});
-  const GameBoard& operator=(const GameBoard& src) = delete;
-  void show() const;
-  void init();
-  bool can_continue() const;
-  void put_dots();
-  void remove_dots();
-private:
-  char to_char(BoardBase::Stone stone) const;
-};
+#include "../include/GameBoard.hpp"
 
 GameBoard::GameBoard(const BoardBase::Position size)
   : BoardBase(size)
@@ -75,5 +57,3 @@ void GameBoard::remove_dots()
 {
   std::replace_if(std::begin(board_), std::end(board_), [](BoardBase::Stone target){ return target == BoardBase::Stone::Dot; }, BoardBase::Stone::Space);
 }
-
-#endif // GAME_BOARD_H_
