+++ src/PlaneVector.cpp
@@ -1,16 +1,4 @@
-#ifndef PLANE_VECTOR_H_
-#define PLANE_VECTOR_H_
-
-struct PlaneVector {
-public:
-  using Point = char;
-  Point x, y;
-  PlaneVector(Point init_x, Point init_y);
-  PlaneVector operator+(const PlaneVector& src);
-  PlaneVector operator*(int n);
-  const PlaneVector& operator+=(const PlaneVector& src);
-  const PlaneVector& operator--();
-};
+#include "../include/PlaneVector.hpp"
 
 PlaneVector::PlaneVector(Point init_x, Point init_y)
   : x{init_x},
@@ -39,5 +27,3 @@ const PlaneVector& PlaneVector::operator--()
   --x; --y;
   return *this;
 }
-
-#endif // PLANE_VECTOR_H_
